+++ test/init_template_test.dart
@@ -19,6 +19,7 @@ const Map<String, dynamic> _schema = {
       'legacy_padding': null,
       'round': null,
       'play_store': null,
+      'image_format': null,
       'image': null,
       'effect': null,
       'adaptive': {